+++ galaxy_save_data/src/chunks/config/misc.rs
@@ -2,7 +2,7 @@
 
 use bilge::prelude::*;
 use binrw::binrw;
-use galaxy_save_core::{bin::Chunk, hash::HashCode, time::OSTime};
+use galaxy_save_core::{bin::Chunk, hash::HashCode, time::Time};
 
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
@@ -17,7 +17,7 @@ pub struct ConfigDataMisc {
     pub flag: ConfigDataMiscFlag,
 
     /// The timestamp representing when the user file was most recently saved.
-    pub last_modified: OSTime,
+    pub last_modified: Time,
 }
 
 impl Chunk for ConfigDataMisc {