+++ galaxy_save_data/src/chunks/sysconf/sys_config.rs
@@ -4,7 +4,7 @@ use binrw::binrw;
 use galaxy_save_core::{
     bin::{BinaryDataContentHeaderSerializer, Chunk, HeaderSerializer},
     hash::HashCode,
-    time::OSTime,
+    time::Time,
 };
 
 #[cfg(feature = "serde")]
@@ -22,11 +22,11 @@ pub struct SysConfigData {
 
     /// The timestamp representing when the player was encouraged to change their TV Type from 50 Hz to 60 Hz.
     #[header_serializer(name = "mTimeAnnounced")]
-    pub time_announced: OSTime,
+    pub time_announced: Time,
 
     /// The timestamp representing when the most recent message was sent to the Wii Message Board.
     #[header_serializer(name = "mTimeSent")]
-    pub time_sent: OSTime,
+    pub time_sent: Time,
 
     /// The number of bytes sent to the Wii Message Board from the date represented in [`time_sent`](#structfield.time_sent).
     #[header_serializer(name = "mSentBytes")]