+++ galaxy_save_core/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Bump `serde` to 1.0.228.
 - Bump `thiserror` to 2.0.17.
+- Rename `OSTime` to `Time`.
 
 ## [0.1.0] - 2025-08-20